+++ src/components/dashboard/DetailDelete.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import { useParams, useRouter } from 'next/navigation';
-import { useRemoveDashboard } from '@/apis/dashboards/queries';
-import Button from '@/components/ui/Button';
 import useAlert from '@/hooks/useAlert';
-import { getErrorMessage } from '@/utils/errorMessage';
 import useConfirm from '@/hooks/useConfirm';
+import { useRemoveDashboard } from '@/apis/dashboards/queries';
+import { getErrorMessage } from '@/utils/errorMessage';
+import Button from '@/components/ui/Button';
 
 export default function DetailDelete() {
   const { id } = useParams<{ id: string }>();
