+++ src/components/dashboard/DetailMembers.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import { isAxiosError } from 'axios';
 import { useState } from 'react';
 import { useParams } from 'next/navigation';
+import { isAxiosError } from 'axios';
 import useAlert from '@/hooks/useAlert';
 import useConfirm from '@/hooks/useConfirm';
 import { useMembersQuery, useRemoveMember } from '@/apis/members/queries';
