+++ src/components/dashboard/header/Profile.tsx
@@ -3,10 +3,10 @@
 import { useEffect, useRef, useState } from 'react';
 import Link from 'next/link';
 import { motion } from 'motion/react';
-import { useGetUser } from '@/apis/users/queries';
-import Avatar from '@/components/ui/Avatar';
 import useAlert from '@/hooks/useAlert';
+import { useGetUser } from '@/apis/users/queries';
 import { useLogout } from '@/apis/auth/queries';
+import Avatar from '@/components/ui/Avatar';
 
 export default function Profile() {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
