+++ src/components/todo/DetailTodo.tsx
@@ -3,17 +3,17 @@
 import { forwardRef } from 'react';
 import Image from 'next/image';
 import useAlert from '@/hooks/useAlert';
+import useConfirm from '@/hooks/useConfirm';
 import { Card } from '@/apis/cards/types';
 import { useColumnsQuery } from '@/apis/columns/queries';
 import { useRemoveCard } from '@/apis/cards/queries';
-import useConfirm from '@/hooks/useConfirm';
-import CommentSection from '@/components/dashboard/CommentSection';
+import { getErrorMessage } from '@/utils/errorMessage';
+import { formatDate } from '@/utils/formatDate';
+import CommentSection from '@/components/todo/CommentSection';
 import { Modal, ModalContent, ModalFooter, ModalHandle, ModalHeader } from '@/components/ui/Modal';
 import Avatar from '@/components/ui/Avatar';
 import TagChip from '@/components/ui/Chip/TagChip';
 import RoundChip from '@/components/ui/Chip/RoundChip';
-import { getErrorMessage } from '@/utils/errorMessage';
-import { formatDate } from '@/utils/formatDate';
 import Dropdown from '@/components/ui/Dropdown';
 import { DEFAULT_CARD_IMAGE_URL } from '@/constants/paths';
 import x_img from '@/assets/icons/x.svg';
